[SWDEV-499995] ESMI Build/Compiler warnings messages (#105)
* [SWDEV-499995] ESMI Build/Compiler warnings messages


[ROCm/amdsmi commit: 2c591ffcc1cc86dae1f51e3fdc43acf65c28c2df]

+++ include/amd_smi/amdsmi.h
@@ -251,8 +251,8 @@ typedef void *amdsmi_cpusocket_handle;
  * @cond @tag{cpu_bm} @endcond
  */
 typedef struct {
-        uint8_t major;      //!< Major version number
-        uint8_t minor;      //!< Minor version number
+        uint32_t major;      //!< Major version number
+        uint32_t minor;      //!< Minor version number
 } amdsmi_hsmp_driver_version_t;
 
 #endif
